Fix weird bug in CMake exporting
