+++ arf/mul_rnd_any.c
@@ -11,16 +11,6 @@
 
 #include "arf.h"
 
-TLS_PREFIX mp_ptr __arf_mul_tmp = NULL;
-TLS_PREFIX slong __arf_mul_alloc = 0;
-
-void _arf_mul_tmp_cleanup(void)
-{
-    flint_free(__arf_mul_tmp);
-    __arf_mul_tmp = NULL;
-    __arf_mul_alloc = 0;
-}
-
 int
 arf_mul_rnd_any(arf_ptr z, arf_srcptr x, arf_srcptr y,
         slong prec, arf_rnd_t rnd)
